resolve conflicts. But Listeners doesn't work properly!!
