+++ src/main/java/faang/school/achievement/event/album/AlbumCreatedEvent.java
@@ -1,4 +1,4 @@
-package faang.school.achievement.event;
+package faang.school.achievement.event.album;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
