--- src/main/java/faang/school/achievement/event_handler/EventHandler.java
@@ -1,5 +1,0 @@
-package faang.school.achievement.event_handler;
-
-public interface EventHandler<T> {
-    void handleEvent(T event);
-}
