+++ src/main/java/faang/school/achievement/handler/EventHandler.java
@@ -1,4 +1,5 @@
 package faang.school.achievement.handler;
+
 public interface EventHandler<T> {
     void handleEvent(T event);
 }
