+++ src/main/java/faang/school/achievement/handler/impl/album/AbstractAchievementHandler.java
@@ -1,11 +1,12 @@
-package faang.school.achievement.handler;
+package faang.school.achievement.handler.impl.album;
 
+import faang.school.achievement.handler.EventHandler;
 import faang.school.achievement.model.Achievement;
 import faang.school.achievement.model.AchievementProgress;
-import faang.school.achievement.event.AlbumCreatedEvent;
+import faang.school.achievement.event.album.AlbumCreatedEvent;
 import faang.school.achievement.model.UserAchievement;
-import faang.school.achievement.service.AchievementCache;
-import faang.school.achievement.service.AchievementService;
+import faang.school.achievement.service.achievement.AchievementCache;
+import faang.school.achievement.service.achievement.AchievementService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -35,5 +36,4 @@ public abstract class AbstractAchievementHandler implements EventHandler<AlbumCr
             }
         }
     }
-
 }
