+++ src/main/java/faang/school/achievement/handler/impl/album/LibrarianAchievementHandler.java
@@ -1,12 +1,14 @@
-package faang.school.achievement.handler;
+package faang.school.achievement.handler.impl.album;
 
-import faang.school.achievement.event.AlbumCreatedEvent;
-import faang.school.achievement.service.AchievementCache;
-import faang.school.achievement.service.AchievementService;
+import faang.school.achievement.event.album.AlbumCreatedEvent;
+import faang.school.achievement.service.achievement.AchievementCache;
+import faang.school.achievement.service.achievement.AchievementService;
 import org.springframework.stereotype.Service;
 
 @Service
 public class LibrarianAchievementHandler extends AbstractAchievementHandler {
+
+    private final String ACHIEVEMENT_TITLE = "LIBRARIAN";
 
     public LibrarianAchievementHandler(AchievementCache achievementCache,
                                        AchievementService achievementService) {
@@ -15,6 +17,6 @@ public class LibrarianAchievementHandler extends AbstractAchievementHandler {
 
     @Override
     public void handleEvent(AlbumCreatedEvent event) {
-        handleAchievement(event, "LIBRARIAN");
+        handleAchievement(event, ACHIEVEMENT_TITLE);
     }
 }
